requirements.txt
gradio
pillow
python-dotenv
requests
transformers
torch
diffusers